Extract classes tests from backend
